делит скрипты на модули
